Improved item value decoding

+++ auth.go
@@ -30,7 +30,7 @@ func AuthWithUserCredentials(clientId string, clientSecret string, username stri
 	return authRequest(data)
 }
 
-func AuthWithAppCredentials(clientId, clientSecret string, appId uint, appToken string) (*AuthToken, error) {
+func AuthWithAppCredentials(clientId, clientSecret string, appId int64, appToken string) (*AuthToken, error) {
 	data := url.Values{
 		"grant_type":    {"app"},
 		"app_id":        {fmt.Sprintf("%d", appId)},
